Injects custom config for live objects into sidebar expression list
